renamed default_settings to defaults
