+++ riskslim/lattice_cpa.py
@@ -3,7 +3,7 @@ import numpy as np
 from cplex.callbacks import HeuristicCallback, LazyConstraintCallback
 from cplex.exceptions import CplexError
 from .bound_tightening import chained_updates
-from .default_settings import DEFAULT_LCPA_SETTINGS
+from .defaults import DEFAULT_LCPA_SETTINGS
 from .helper_functions import cast_to_integer, is_integer, print_log, validate_settings
 from .heuristics import discrete_descent, sequential_rounding
 from .initialization import initialize_lattice_cpa
